requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
requests
